pyproject.toml
[project]
name = "backend-sibp"
version = "0.1.0"
description = ""
authors = [
    {name = "Backend SEED",email = "backend@seedtec.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "cryptography (>=45.0.0,<46.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
